requirements.txt
crewai
crewai-tools
python-dotenv
duckduckgo-search
langchain-community
beautifulsoup4
